+++ src/main/kotlin/stringcalculator/enums/Operator.kt
@@ -4,15 +4,15 @@ import stringcalculator.constants.ErrorType
 
 enum class Operator(
     private val symbol: String,
-    private val operate: (Int, Int) -> Int
+    private val operate: (Double, Double) -> Double
 ) {
-    ADD("+", {num1: Int, num2: Int -> num1 + num2}),
-    SUBTRACT("-", { num1: Int, num2: Int -> num1 - num2}),
-    MULTIPLE("*", { num1: Int, num2: Int -> num1 * num2}),
-    DIVIDE("/", { num1: Int, num2: Int -> num1 / num2})
+    ADD("+", { num1: Double, num2: Double -> num1 + num2}),
+    SUBTRACT("-", { num1: Double, num2: Double -> num1 - num2}),
+    MULTIPLE("*", { num1: Double, num2: Double -> num1 * num2}),
+    DIVIDE("/", { num1: Double, num2: Double -> num1 / num2})
     ;
 
-    fun calculate(num1: Int, num2: Int): Int = operate(num1, num2)
+    fun calculate(num1: Double, num2: Double): Double = operate(num1, num2)
 
     companion object {
         fun of(input: String): Operator {
